LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-getComputedStyle-002-expected.txt

Lorem ipsum dolor sit amet.
Lorem ipsum dolor sit amet.

<<<<<<< HEAD
FAIL getComputedStyle() with fragmented containing block in multicolumn layout assert_equals: expected "25px" but got "135px"
FAIL getComputedStyle() with fragmented containing block in inline layout assert_equals: expected "-120px" but got "0px"
=======
FAIL getComputedStyle() with fragmented containing block in multicolumn layout assert_equals: expected "-85px" but got "25px"
PASS getComputedStyle() with fragmented containing block in inline layout
>>>>>>> 90aaff9c95eb ([css-anchor-position-1] Integrate anchor positioning with multicol layout)
